Refactor world model architecture to eliminate U-Net skips and enhance embedding supervision
This commit updates the world model configuration and architecture by removing the U-Net skip connections from the `PerceptionAutoencoder`, transitioning to a skip-free `Encoder`. The `embed_decoder_bottleneck` has been integrated as a direct supervision mechanism for the embedding, ensuring that the RSSM receives a richer signal. The training loss structure has been adjusted to reflect these changes, with `recon_bottleneck_scale` set to zero to prevent double counting. Additionally, the configuration file has been modified to align with these architectural updates, ensuring that old checkpoints will not load and requiring a fresh start with `RESUME = None`. This refactor aims to improve the model's ability to capture fine details and enhance overall training performance.

+++ src/models/rssm.py
@@ -93,7 +93,7 @@ def sample_onehot_ste(probs: Tensor) -> Tensor:
         One-hot samples `[..., classes]` (STE: forward = hard sample,
         backward = gradient flows as if the output were `probs` directly).
     """
-    flat = probs.reshape(-1, probs.shape[-1])
+    flat = probs.reshape(-1, probs.shape[-1]).float()
     indices = torch.multinomial(flat, num_samples=1).squeeze(-1)
     hard = F.one_hot(indices, num_classes=probs.shape[-1]).float()
     hard = hard.view_as(probs)
